Relocate serialization functions to correpsonding column tests
