+++ python/cudf/cudf/tests/test_serialize.py
@@ -9,7 +9,7 @@ import pytest
 
 import cudf
 from cudf.tests import utils
-from cudf.tests.utils import _decimal_series, assert_eq
+from cudf.tests.utils import assert_eq
 
 
 @pytest.mark.parametrize(
@@ -267,86 +267,6 @@ def test_serialize_string_check_buffer_sizes():
     header, frames = df.serialize()
     got = sum(b.nbytes for b in frames)
     assert expect == got
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": cudf.Series(["a", "b", "c", "a", "c", "b"]).astype("category")},
-        {
-            "a": cudf.Series(["a", "a", "b", "b"]).astype("category"),
-            "b": cudf.Series(["b", "b", "c", "c"]).astype("category"),
-            "c": cudf.Series(["c", "c", "a", "a"]).astype("category"),
-        },
-        {
-            "a": cudf.Series(["a", None, "b", "b"]).astype("category"),
-            "b": cudf.Series(["b", "b", None, "c"]).astype("category"),
-            "c": cudf.Series(["c", "c", "a", None]).astype("category"),
-        },
-    ],
-)
-def test_serialize_categorical_columns(data):
-    df = cudf.DataFrame(data)
-    recreated = df.__class__.deserialize(*df.serialize())
-    assert_eq(recreated, df)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {"a": [[]]},
-        {"a": [[1, 2, None, 4]]},
-        {"a": [["cat", None, "dog"]]},
-        {
-            "a": [[1, 2, 3, None], [4, None, 5]],
-            "b": [None, ["fish", "bird"]],
-            "c": [[], []],
-        },
-        {"a": [[1, 2, 3, None], [4, None, 5], None, [6, 7]]},
-    ],
-)
-def test_serialize_list_columns(data):
-    df = cudf.DataFrame(data)
-    recreated = df.__class__.deserialize(*df.serialize())
-    assert_eq(recreated, df)
-
-
-@pytest.mark.parametrize(
-    "data",
-    [
-        {
-            "a": _decimal_series(
-                ["1", "2", "3"], dtype=cudf.Decimal64Dtype(1, 0)
-            )
-        },
-        {
-            "a": _decimal_series(
-                ["1", "2", "3"], dtype=cudf.Decimal64Dtype(1, 0)
-            ),
-            "b": _decimal_series(
-                ["1.0", "2.0", "3.0"], dtype=cudf.Decimal64Dtype(2, 1)
-            ),
-            "c": _decimal_series(
-                ["10.1", "20.2", "30.3"], dtype=cudf.Decimal64Dtype(3, 1)
-            ),
-        },
-        {
-            "a": _decimal_series(
-                ["1", None, "3"], dtype=cudf.Decimal64Dtype(1, 0)
-            ),
-            "b": _decimal_series(
-                ["1.0", "2.0", None], dtype=cudf.Decimal64Dtype(2, 1)
-            ),
-            "c": _decimal_series(
-                [None, "20.2", "30.3"], dtype=cudf.Decimal64Dtype(3, 1)
-            ),
-        },
-    ],
-)
-def test_serialize_decimal_columns(data):
-    df = cudf.DataFrame(data)
-    recreated = df.__class__.deserialize(*df.serialize())
-    assert_eq(recreated, df)
 
 
 def test_deserialize_cudf_0_16(datadir):
